requirements.txt
tensorflow==2.15.0
streamlit==1.35.0
Pillow==10.1.0
scikit-image==0.22.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
protobuf==3.20.3
source venv/Scripts/activate
